Refactor

+++ src/main/java/ru/hse/lorddux/queue/DeleteQueueMessagesClient.java
@@ -4,7 +4,7 @@ import com.microsoft.azure.storage.StorageException;
 import com.microsoft.azure.storage.queue.CloudQueueMessage;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import ru.hse.lorddux.executor.PythonExecutor;
+import ru.hse.lorddux.executors.PythonExecutor;
 import ru.hse.lorddux.utils.ExecutorQueuePool;
 
 import java.util.Collection;
@@ -27,8 +27,11 @@ public class DeleteQueueMessagesClient implements Runnable {
 
     @Override
     public void run() {
+        CloudQueueMessage messageToDelete;
         while (!stopFlag) {
-            CloudQueueMessage messageToDelete = queuePool.poll(1000L);
+            if ((messageToDelete = queuePool.poll(1000L)) == null) {
+                return;
+            }
             try {
                 log_.debug(String.format("Delete message %s", messageToDelete.getId()));
                 queueProcessor.deleteTask(messageToDelete);
